+++ apps/hash-frontend/src/pages/flows.page/flow-definition/run-flow-modal/types.ts
@@ -1,19 +1,63 @@
+import type { Subtype } from "@local/advanced-types/subtype";
+import type {
+  FormattedText,
+  GoogleSheet,
+  PayloadKind,
+} from "@local/hash-isomorphic-utils/flows/types";
+import type { ActorTypeDataType } from "@local/hash-isomorphic-utils/system-types/googlesheetsintegration";
 import type {
   Entity,
   EntityTypeWithMetadata,
   OwnedById,
 } from "@local/hash-subgraph";
 
-export type LocalInputValues = {
-  Entity: Entity;
-  Text: string;
-  Number: number;
-  Boolean: boolean;
-  VersionedUrl: EntityTypeWithMetadata;
-  WebId: OwnedById;
-};
+const unsupportedPayloadKinds = [
+  /** @todo support ActorType to vary Sheet output formatting */
+  "ActorType",
+  /** @todo support EntityId – requires updating the EntitySelector */
+  "EntityId",
+  "PersistedEntities",
+  "PersistedEntity",
+  "ProposedEntity",
+  "ProposedEntityWithResolvedLinks",
+  "FormattedText",
+  "WebPage",
+] as const satisfies ReadonlyArray<PayloadKind>;
 
-export type LocalPayloadKind = keyof LocalInputValues;
+type UnsupportedPayloadKind = (typeof unsupportedPayloadKinds)[number];
+
+export type LocalPayloadKind = Exclude<PayloadKind, UnsupportedPayloadKind>;
+
+export const isSupportedPayloadKind = (
+  kind: PayloadKind,
+): kind is LocalPayloadKind =>
+  !unsupportedPayloadKinds.includes(kind as UnsupportedPayloadKind);
+
+export type LocalInputValue =
+  | Entity
+  | EntityTypeWithMetadata
+  | FormattedText
+  | GoogleSheet
+  | OwnedById
+  | string
+  | number
+  | boolean;
+
+export type LocalInputValues = Subtype<
+  Record<LocalPayloadKind, LocalInputValue>,
+  {
+    ActorType: ActorTypeDataType;
+    Entity: Entity;
+    FormattedText: FormattedText;
+    GoogleAccountId: string;
+    GoogleSheet: GoogleSheet;
+    Text: string;
+    Number: number;
+    Boolean: boolean;
+    VersionedUrl: EntityTypeWithMetadata;
+    WebId: OwnedById;
+  }
+>;
 
 export type LocalPayload = {
   [K in LocalPayloadKind]: {
@@ -21,3 +65,10 @@ export type LocalPayload = {
     value?: LocalInputValues[K] | LocalInputValues[K][];
   };
 }[LocalPayloadKind];
+
+export type FormState = {
+  [outputName: string]: {
+    outputName: string;
+    payload: LocalPayload;
+  };
+};
