+++ libs/@local/hash-isomorphic-utils/src/google-integration.ts
@@ -2,21 +2,6 @@ import type { EntityId } from "@local/hash-subgraph";
 
 type ErrorResponse = { error: string };
 
-export type CreateOrUpdateSheetsIntegrationRequest = {
-  audience: "human" | "machine";
-  existingIntegrationEntityId?: EntityId;
-  googleAccountId: string;
-  queryEntityId: EntityId;
-  spreadsheetId?: string;
-  newFileName?: string;
-};
-
-export type CreateOrUpdateSheetsIntegrationResponse =
-  | {
-      integrationEntityId: EntityId;
-    }
-  | ErrorResponse;
-
 export type GoogleOAuth2CallbackRequest = {
   code: string;
 };
